requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
google-generativeai==0.3.2
python-docx==1.1.0
PyPDF2==3.0.1
pytz==2023.3
psutil==5.9.6
pydantic==2.5.0